third_party/blink/web_tests/fast/xsl/xslt/default-html-quirk-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: XSLTProcessor is not defined
Test for bug 11694: XSLT output method does not default to HTML when the target document is HTML.

Should be green:
Should be black:
